requirements.txt
streamlit>=1.36.0
langchain>=0.2.0
langchain-groq>=0.1.3
PyPDF2>=3.0.1
python-docx>=1.1.0
pydantic>=2.6.0
